demo-hack/My_AI_Employee/pyproject.toml
[project]
name = "my-ai-employee"
version = "0.2.0"
description = "Silver Tier AI Employee - Multi-channel monitoring with HITL approval workflow"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Bronze Tier dependencies
    "pytest>=9.0.2",
    "python-frontmatter>=1.1.0",
    "python-dotenv>=1.0.0",
    "watchdog>=6.0.0",
    # Silver Tier dependencies
    "fastmcp>=0.1.0",
    "playwright>=1.40.0",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.1",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["watchers*", "utils*", "vault_ops*", "triage*"]
exclude = ["AI_Employee_Vault*", "test_watch_folder*", "tests*"]
